+++ internal/search/run/run.go
@@ -4,11 +4,13 @@ import (
 	"context"
 	"fmt"
 	"math"
+	"time"
 
 	"github.com/sourcegraph/sourcegraph/internal/errcode"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/search"
 	"github.com/sourcegraph/sourcegraph/internal/search/query"
+	searchrepos "github.com/sourcegraph/sourcegraph/internal/search/repos"
 	"github.com/sourcegraph/sourcegraph/internal/search/streaming"
 	"github.com/sourcegraph/sourcegraph/internal/vcs"
 	"github.com/sourcegraph/sourcegraph/schema"
@@ -52,6 +54,27 @@ func (inputs SearchInputs) MaxResults() int {
 	}
 
 	return defaultMaxSearchResults
+}
+
+// Timeout computes the timeout for the query.
+func (inputs SearchInputs) Timeout() time.Duration {
+	// The default timeout to use for queries.
+	d := 20 * time.Second
+
+	maxTimeout := time.Duration(searchrepos.SearchLimits().MaxTimeoutSeconds) * time.Second
+
+	if timeout := inputs.Query.Timeout(); timeout != nil {
+		d = *timeout
+	} else if inputs.Query.Count() != nil {
+		// If `count:` is set but `timeout:` is not explicitly set, use the max timeout
+		d = maxTimeout
+	}
+
+	if d > maxTimeout {
+		d = maxTimeout
+	}
+
+	return d
 }
 
 // SearchPaginationInfo describes information around a paginated search